+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * `appInfo` dumped app argument
   * `ssh` dumped app argument
   * `tunnel` dumped app argument
+  * `build` dumped app result field
   * naming scheme for context directories changed from app:tag to tag (current build is now _)
   * build now creates a repository tag without app name
   * dumped container label `app`